Fix warnings, mostly about implicit int conversions
This should fix all non-erros, leaving only warnings which
are indeed potential problems.

The following warnings should be safe to ignore:

* Conversions of various offsets from uint64 to
  std::streamoff/int64 are safe because such offsets have
  been obtained via tellg() and other functions
  returning std::streamoff in the first place.
* It also works vice-versa since tellg() should not
  return negative offsets with exceptions enabled.
* Conversions from char to unsigned char are also ok.
* Unused diag arguments can be ignored (those might be
  useful later).
* Annotate all intended fallthoughs.

+++ matroska/matroskaattachment.cpp
@@ -173,7 +173,7 @@ void MatroskaAttachmentMaker::make(ostream &stream, Diagnostics &diag) const
     if (attachment().data() && attachment().data()->size()) {
         BE::getBytes(static_cast<uint16>(MatroskaIds::FileData), buff);
         stream.write(buff, 2);
-        len = EbmlElement::makeSizeDenotation(attachment().data()->size(), buff);
+        len = EbmlElement::makeSizeDenotation(static_cast<uint64>(attachment().data()->size()), buff);
         stream.write(buff, len);
         attachment().data()->copyTo(stream);
     }
